requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cryptacular=1.6.2=pypi_0
cryptography=41.0.3=py39hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
defusedxml=0.7.1=pypi_0
easydict=1.11=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
focalloss=1.0.0=dev_0
fonttools=4.46.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
ftfy=6.1.3=pypi_0
fvcore=0.1.5.post20221221=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
greenlet=3.0.3=pypi_0
huggingface-hub=0.20.1=pypi_0
hupper=1.12=pypi_0
idna=3.4=py39h06a4308_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.1.2=0
libcurand=10.3.4.101=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.4=pypi_0
matplotlib=3.8.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nltk=3.8.1=pypi_0
numpy=1.26.2=py39h5f9d8c6_0
numpy-base=1.26.2=py39hb5e798b_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pastedeploy=3.1.0=pypi_0
pbkdf2=1.3=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=py39h06a4308_0
plaster=1.1.2=pypi_0
plaster-pastedeploy=1.0.1=pypi_0
portalocker=2.8.2=pypi_0
prettytable=3.9.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pyramid=2.0.2=pypi_0
pyramid-mailer=0.15.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python3-openid=3.2.0=pypi_0
pytorch=1.13.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytorch-warmup-scheduler=0.0.1=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h5eee18b_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.4.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.2=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
torchaudio=0.13.1=py39_cu117
torchvision=0.14.1=py39_cu117
tqdm=4.66.1=pypi_0
transaction=4.0=pypi_0
transformers=4.36.2=pypi_0
translationstring=1.4=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023.4=pypi_0
urllib3=1.26.18=py39h06a4308_0
velruse=1.1.1=pypi_0
venusian=3.1.0=pypi_0
wcwidth=0.2.12=pypi_0
webob=1.8.7=pypi_0
wheel=0.41.2=py39h06a4308_0
wtforms=3.1.2=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zope-deprecation=5.0=pypi_0
zope-interface=6.1=pypi_0
zope-sqlalchemy=3.1=pypi_0
zstd=1.5.5=hc292b87_0
